+++ processor/src/main/java/react4j/processor/Constants.java
@@ -15,6 +15,8 @@ final class Constants
   @Nonnull
   static final String INPUT_CLASSNAME = "react4j.annotations.Input";
   @Nonnull
+  static final String PUBLISH_CLASSNAME = "react4j.annotations.Publish";
+  @Nonnull
   static final String SUPPRESS_REACT4J_WARNINGS_CLASSNAME = "react4j.annotations.SuppressReact4jWarnings";
   @Nonnull
   static final String INPUT_REF_CLASSNAME = "react4j.annotations.InputRef";