+++ processor/src/test/fixtures/expected/com/example/publish/React4j_BasicPublishView.java
@@ -1,0 +1,124 @@
+package com.example.publish;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import java.util.Objects;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.Context;
+import react4j.Contexts;
+import react4j.React;
+import react4j.ReactNode;
+import react4j.internal.OnComponentWillUnmount;
+import react4j.internal.OnShouldComponentUpdate;
+import react4j.internal.ViewConstructorFunction;
+
+@SuppressWarnings("Arez:UnnecessaryAllowEmpty")
+@ArezComponent(
+    name = "BasicPublishView",
+    disposeNotifier = Feature.DISABLE,
+    dagger = Feature.DISABLE,
+    sting = Feature.DISABLE,
+    allowEmpty = true
+)
+@Generated("react4j.processor.React4jProcessor")
+abstract class React4j_BasicPublishView extends BasicPublishView {
+  @Nonnull
+  private final react4j.internal.NativeView $$react4j$$_nativeView;
+
+  React4j_BasicPublishView(@Nonnull final react4j.internal.NativeView $$react4j$$_nativeView) {
+    this.$$react4j$$_nativeView = Objects.requireNonNull( $$react4j$$_nativeView );
+  }
+
+  @Nonnull
+  private static ViewConstructorFunction getConstructorFunction() {
+    final ViewConstructorFunction viewConstructor = ( React.shouldStoreDebugDataAsState() || React.shouldValidateInputValues() ) ? NativeView::new : LiteNativeView::new;
+    if ( React.enableViewNames() ) {
+      Js.asPropertyMap( viewConstructor ).set( "displayName", "BasicPublishView" );
+    }
+    return viewConstructor;
+  }
+
+  private boolean $$react4j$$_shouldComponentUpdate(
+      @Nullable final JsPropertyMap<Object> nextInputs) {
+    assert null != nextInputs;
+    return false;
+  }
+
+  private void $$react4j$$_componentWillUnmount() {
+    ((Arez_React4j_BasicPublishView) this).dispose();
+  }
+
+  @Nullable
+  ReactNode $$react4j$$_render() {
+    assert Disposable.isNotDisposed( this );
+    return ContextHolder.CONTEXT_getMyContextVar.provide( getMyContextVar(), render() );
+  }
+
+  private static final class ContextHolder {
+    @Nonnull
+    private static final Context<String> CONTEXT_getMyContextVar = Contexts.get( String.class );
+
+    private ContextHolder() {
+    }
+  }
+
+  static final class Factory {
+    @Nonnull
+    static final ViewConstructorFunction TYPE = getConstructorFunction();
+  }
+
+  private static final class LiteNativeView extends react4j.internal.NativeView implements OnShouldComponentUpdate {
+    @Nonnull
+    private final React4j_BasicPublishView $$react4j$$_view;
+
+    @JsConstructor
+    LiteNativeView(@Nullable final JsPropertyMap<Object> inputs) {
+      super( inputs );
+      $$react4j$$_view = new Arez_React4j_BasicPublishView( this );
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextInputs) {
+      return $$react4j$$_view.$$react4j$$_shouldComponentUpdate( nextInputs );
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_view.$$react4j$$_render();
+    }
+  }
+
+  private static final class NativeView extends react4j.internal.NativeView implements OnShouldComponentUpdate, OnComponentWillUnmount {
+    @Nonnull
+    private final React4j_BasicPublishView $$react4j$$_view;
+
+    @JsConstructor
+    NativeView(@Nullable final JsPropertyMap<Object> inputs) {
+      super( inputs );
+      $$react4j$$_view = new Arez_React4j_BasicPublishView( this );
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextInputs) {
+      return $$react4j$$_view.$$react4j$$_shouldComponentUpdate( nextInputs );
+    }
+
+    @Override
+    public final void componentWillUnmount() {
+      $$react4j$$_view.$$react4j$$_componentWillUnmount();
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_view.$$react4j$$_render();
+    }
+  }
+}