+++ processor/src/test/fixtures/input/com/example/publish/BasicPublishView.java
@@ -1,0 +1,22 @@
+package com.example.publish;
+
+import react4j.ReactNode;
+import react4j.annotations.Publish;
+import react4j.annotations.Render;
+import react4j.annotations.View;
+
+@View
+abstract class BasicPublishView
+{
+  @Publish
+  String getMyContextVar()
+  {
+    return "";
+  }
+
+  @Render
+  ReactNode render()
+  {
+    return null;
+  }
+}