+++ processor/src/test/fixtures/input/com/example/publish/ProtectedAccessFromBasePublishModel.java
@@ -1,0 +1,10 @@
+package com.example.publish;
+
+import com.example.publish.other.BaseProtectedAccessPublishModel;
+import react4j.annotations.View;
+
+@View
+abstract class ProtectedAccessFromBasePublishModel
+  extends BaseProtectedAccessPublishModel
+{
+}